✨ feat(web): create channels in place on the agent profile
添加渠道 opens a ChannelCreateSheet with the agent fixed instead of
jumping to settings; NewChannelForm is extracted and shared, and the
feishu/weixin scan Dialog becomes an in-surface covering step — it was
already an illegal Dialog-in-Sheet on the settings page.
